+++ src/server/API/appsettings.json
@@ -45,5 +45,8 @@
   "SerializationSettings": {
     "UseSystemTextJson": true,
     "UseNewtonsoftJson": false
+  },
+  "ApplicationSettings": {
+    "ApiUrl": "https://localhost:5001/"
   }
 }
