+++ src/server/Modules/Catalog/Modules.Catalog.Core/Features/Products/Queries/ProductQueryHandler.cs
@@ -18,11 +18,13 @@ using FluentPOS.Modules.Catalog.Core.Exceptions;
 using FluentPOS.Shared.Core.Constants;
 using FluentPOS.Shared.Core.Extensions;
 using FluentPOS.Shared.Core.Mappings.Converters;
+using FluentPOS.Shared.Core.Settings;
 using FluentPOS.Shared.Core.Wrapper;
 using FluentPOS.Shared.DTOs.Catalogs.Products;
 using MediatR;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.Localization;
+using Microsoft.Extensions.Options;
 
 namespace FluentPOS.Modules.Catalog.Core.Features.Products.Queries
 {
@@ -33,12 +35,16 @@ namespace FluentPOS.Modules.Catalog.Core.Features.Products.Queries
     {
         private readonly ICatalogDbContext _context;
         private readonly IMapper _mapper;
+        private readonly ApplicationSettings _applicationSettings;
         private readonly IStringLocalizer<ProductQueryHandler> _localizer;
 
-        public ProductQueryHandler(ICatalogDbContext context, IMapper mapper, IStringLocalizer<ProductQueryHandler> localizer)
+        public ProductQueryHandler(ICatalogDbContext context, IMapper mapper,
+            ApplicationSettings applicationSettings,
+            IStringLocalizer<ProductQueryHandler> localizer)
         {
             _context = context;
             _mapper = mapper;
+            _applicationSettings = applicationSettings;
             _localizer = localizer;
         }
 
@@ -99,11 +105,12 @@ namespace FluentPOS.Modules.Catalog.Core.Features.Products.Queries
         }
 
 #pragma warning disable RCS1046 // Asynchronous method name should end with 'Async'.
+
         public async Task<Result<string>> Handle(GetProductImageQuery query, CancellationToken cancellationToken)
 #pragma warning restore RCS1046 // Asynchronous method name should end with 'Async'.
         {
             string data = await _context.Products.Where(p => p.Id == query.Id).Select(x => x.ImageUrl).FirstOrDefaultAsync(cancellationToken);
-            return await Result<string>.SuccessAsync(data: AppConstants.ApiUrl + data);
+            return await Result<string>.SuccessAsync(data: _applicationSettings.ApiUrl + data);
         }
     }
 }
