+++ src/server/Shared/Shared.Core/Settings/ApplicationSettings.cs
@@ -1,15 +1,15 @@
 // --------------------------------------------------------------------------------------------------
-// <copyright file="AppConstants.cs" company="FluentPOS">
+// <copyright file="ApplicationSettings.cs" company="FluentPOS">
 // Copyright (c) FluentPOS. All rights reserved.
 // The core team: Mukesh Murugan (iammukeshm), Chhin Sras (chhinsras), Nikolay Chebotov (unchase).
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 // </copyright>
 // --------------------------------------------------------------------------------------------------
 
-namespace FluentPOS.Shared.Core.Constants
+namespace FluentPOS.Shared.Core.Settings
 {
-    public static class AppConstants
+    public class ApplicationSettings
     {
-        public const string ApiUrl = "https://localhost:5001/";
+        public string ApiUrl { get; set; }
     }
 }
